test(05-14): rename and split BugEvidenceTest into topic-named classes (D-10)
BugEvidenceTest's disposition is resolved as FIXED: the class was 15/15
green and its assertions already asserted correct behaviour, but its name
and class-level javadoc still claimed the tests were "written to FAIL with
the current implementation" (PR #130 evidence). That framing is what misled
three control documents into stating BugEvidenceTest fails on purpose.

Split the six former @Nested groups into three topic-named classes per
D-10's three named topics, with every assertion carried across unchanged:

- core/ElementReconciliationTest (8 tests) — CanUpdateBug, ReconciliationBug
  and CombinedBugs, which all assert Element.canUpdate()/updateChild()'s
  type-mismatch reconciliation contract (CombinedBugs via ContainerElement
  as harness).
- widgets/ContainerGridViewChildTrackingTest (6 tests) — ContainerChildTrackingBug
  and GridViewChildTrackingBug, asserting getChildren() tracks additions/
  removals/unmounts after Element.updateChildren() (converged onto one
  shared implementation by plan 05-13).
- widgets/ItemDisplayFieldShadowingTest (1 test) — ShadowedFieldFixed.

Pre-split baseline (re-measured, not assumed): mvn -o test -Dtest=BugEvidenceTest
— 15 tests, 0 failures. Post-split: 15 tests, 0 failures, identical individual
test names (module-private @DisplayName reworded off "bug #N" framing where it
said so; kept where already behavioural). No new class carries @Disabled or
@Tag. No test found redundant.

(cherry picked from commit 7ede5eec04804ce733edfd141d777cef5fc59c49)
